requirements.txt
beautifulsoup4==4.8.1
newspaper3k==0.2.8
rouge==0.3.1